Refactor OSRS tests into modular suite and update tracker
