+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -9,8 +9,9 @@
 - **Current Compliance**: 72.9% (433/594 files)
 - **Target Compliance**: 100% (594/594 files)
 - **Status**: 🟢 **EXCELLENT PROGRESS - MAJOR MILESTONE ACHIEVED**
-- **Last Updated**: 2025-08-24
+- **Last Updated**: 2025-08-25
 - **Progress**: 161 files remaining (down from 208)
+- **CRIT-004**: ✅ Completed
 - **Major Achievement**: 🎉 **ZERO files over 800 lines remaining!**
 
 > **🎯 NEW FOCUS: We are now ONLY targeting files with 400+ lines for modularization.**
