+++ tests/gaming/osrs/test_osrs_game_setup.py
