+++ tests/gaming/osrs/test_osrs_scenarios.py
