+++ tests/gaming/osrs/test_osrs_validation.py
